% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{batch_get_application_revisions}
\alias{batch_get_application_revisions}
\title{Gets information about one or more application revisions}
\usage{
batch_get_application_revisions(applicationName, revisions)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application about which to get revision information.}

\item{revisions}{[required] Information to get about the application revisions, including type and location.}
}
\description{
Gets information about one or more application revisions.
}
\section{Accepted Parameters}{
\preformatted{batch_get_application_revisions(
  applicationName = "string",
  revisions = list(
    list(
      revisionType = "S3"|"GitHub"|"String"|"AppSpecContent",
      s3Location = list(
        bucket = "string",
        key = "string",
        bundleType = "tar"|"tgz"|"zip"|"YAML"|"JSON",
        version = "string",
        eTag = "string"
      ),
      gitHubLocation = list(
        repository = "string",
        commitId = "string"
      ),
      string = list(
        content = "string",
        sha256 = "string"
      ),
      appSpecContent = list(
        content = "string",
        sha256 = "string"
      )
    )
  )
)
}
}

