% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{batch_get_applications}
\alias{batch_get_applications}
\title{Gets information about one or more applications}
\usage{
batch_get_applications(applicationNames)
}
\arguments{
\item{applicationNames}{[required] A list of application names separated by spaces.}
}
\description{
Gets information about one or more applications.
}
\section{Accepted Parameters}{
\preformatted{batch_get_applications(
  applicationNames = list(
    "string"
  )
)
}
}

