% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{batch_get_deployment_groups}
\alias{batch_get_deployment_groups}
\title{Gets information about one or more deployment groups}
\usage{
batch_get_deployment_groups(applicationName, deploymentGroupNames)
}
\arguments{
\item{applicationName}{The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}

\item{deploymentGroupNames}{The deployment groups' names.}
}
\description{
Gets information about one or more deployment groups.
}
\examples{

}
