% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{batch_get_deployments}
\alias{batch_get_deployments}
\title{Gets information about one or more deployments}
\usage{
batch_get_deployments(deploymentIds)
}
\arguments{
\item{deploymentIds}{A list of deployment IDs, separated by spaces.}
}
\description{
Gets information about one or more deployments.
}
\examples{

}
