% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{continue_deployment}
\alias{continue_deployment}
\title{For a blue/green deployment, starts the process of rerouting traffic from instances in the original environment to instances in the replacement environment without waiting for a specified wait time to elapse}
\usage{
continue_deployment(deploymentId = NULL)
}
\arguments{
\item{deploymentId}{The deployment ID of the blue/green deployment for which you want to start rerouting traffic to the replacement environment.}
}
\description{
For a blue/green deployment, starts the process of rerouting traffic from instances in the original environment to instances in the replacement environment without waiting for a specified wait time to elapse. (Traffic rerouting, which is achieved by registering instances in the replacement environment with the load balancer, can start as soon as all instances have a status of Ready.)
}
\section{Accepted Parameters}{
\preformatted{continue_deployment(
  deploymentId = "string"
)
}
}

