% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{create_application}
\alias{create_application}
\title{Creates an application}
\usage{
create_application(
  applicationName = "string",
  computePlatform = "Server"|"Lambda"|"ECS"
)
}
\arguments{
\item{applicationName}{[required] The name of the application. This name must be unique with the applicable IAM user or AWS account.}

\item{computePlatform}{The destination platform type for the deployment (\code{Lambda} or \code{Server}).}
}
\description{
Creates an application.
}
\examples{

}
