% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{create_deployment}
\alias{create_deployment}
\title{Deploys an application revision through the specified deployment group}
\usage{
create_deployment(applicationName, deploymentGroupName = NULL,
  revision = NULL, deploymentConfigName = NULL, description = NULL,
  ignoreApplicationStopFailures = NULL, targetInstances = NULL,
  autoRollbackConfiguration = NULL, updateOutdatedInstancesOnly = NULL,
  fileExistsBehavior = NULL)
}
\arguments{
\item{applicationName}{The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}

\item{deploymentGroupName}{The name of the deployment group.}

\item{revision}{The type and location of the revision to deploy.}

\item{deploymentConfigName}{The name of a deployment configuration associated with the applicable IAM user or AWS account.

If not specified, the value configured in the deployment group will be used as the default. If the deployment group does not have a deployment configuration associated with it, then CodeDeployDefault.OneAtATime will be used by default.}

\item{description}{A comment about the deployment.}

\item{ignoreApplicationStopFailures}{If set to true, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will not be considered to have failed at that point and will continue on to the BeforeInstall deployment lifecycle event.

If set to false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to fail to an instance, the deployment to that instance will stop, and the deployment to that instance will be considered to have failed.}

\item{targetInstances}{Information about the instances that will belong to the replacement environment in a blue/green deployment.}

\item{autoRollbackConfiguration}{Configuration information for an automatic rollback that is added when a deployment is created.}

\item{updateOutdatedInstancesOnly}{Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.}

\item{fileExistsBehavior}{Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

The fileExistsBehavior parameter takes any of the following values:
\itemize{
\item DISALLOW: The deployment fails. This is also the default behavior if no option is specified.
\item OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.
\item RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.
}}
}
\description{
Deploys an application revision through the specified deployment group.
}
\examples{

}
