% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{delete_application}
\alias{delete_application}
\title{Deletes an application}
\usage{
delete_application(applicationName)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}
}
\description{
Deletes an application.
}
\section{Accepted Parameters}{
\preformatted{delete_application(
  applicationName = "string"
)
}
}

