% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{delete_deployment_config}
\alias{delete_deployment_config}
\title{Deletes a deployment configuration}
\usage{
delete_deployment_config(deploymentConfigName)
}
\arguments{
\item{deploymentConfigName}{[required] The name of a deployment configuration associated with the applicable IAM user or AWS account.}
}
\description{
Deletes a deployment configuration.
}
\details{
A deployment configuration cannot be deleted if it is currently in use. Predefined configurations cannot be deleted.
}
\section{Accepted Parameters}{
\preformatted{delete_deployment_config(
  deploymentConfigName = "string"
)
}
}

