% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{delete_deployment_group}
\alias{delete_deployment_group}
\title{Deletes a deployment group}
\usage{
delete_deployment_group(applicationName, deploymentGroupName)
}
\arguments{
\item{applicationName}{The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}

\item{deploymentGroupName}{The name of an existing deployment group for the specified application.}
}
\description{
Deletes a deployment group.
}
\examples{

}
