% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{delete_git_hub_account_token}
\alias{delete_git_hub_account_token}
\title{Deletes a GitHub account connection}
\usage{
delete_git_hub_account_token(tokenName = NULL)
}
\arguments{
\item{tokenName}{The name of the GitHub account connection to delete.}
}
\description{
Deletes a GitHub account connection.
}
\section{Accepted Parameters}{
\preformatted{delete_git_hub_account_token(
  tokenName = "string"
)
}
}

