% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{deregister_on_premises_instance}
\alias{deregister_on_premises_instance}
\title{Deregisters an on-premises instance}
\usage{
deregister_on_premises_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the on-premises instance to deregister.}
}
\description{
Deregisters an on-premises instance.
}
\section{Accepted Parameters}{
\preformatted{deregister_on_premises_instance(
  instanceName = "string"
)
}
}

