% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_application}
\alias{get_application}
\title{Gets information about an application}
\usage{
get_application(
  applicationName = "string"
)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}
}
\description{
Gets information about an application.
}
\examples{

}
