% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_application_revision}
\alias{get_application_revision}
\title{Gets information about an application revision}
\usage{
get_application_revision(applicationName, revision)
}
\arguments{
\item{applicationName}{The name of the application that corresponds to the revision.}

\item{revision}{Information about the application revision to get, including type and location.}
}
\description{
Gets information about an application revision.
}
\examples{

}
