% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_deployment_config}
\alias{get_deployment_config}
\title{Gets information about a deployment configuration}
\usage{
get_deployment_config(deploymentConfigName)
}
\arguments{
\item{deploymentConfigName}{[required] The name of a deployment configuration associated with the IAM user or AWS account.}
}
\description{
Gets information about a deployment configuration.
}
\section{Accepted Parameters}{
\preformatted{get_deployment_config(
  deploymentConfigName = "string"
)
}
}

