% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_deployment_instance}
\alias{get_deployment_instance}
\title{Gets information about an instance as part of a deployment}
\usage{
get_deployment_instance(deploymentId, instanceId)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{instanceId}{[required] The unique ID of an instance in the deployment group.}
}
\description{
Gets information about an instance as part of a deployment.
}
\section{Accepted Parameters}{
\preformatted{get_deployment_instance(
  deploymentId = "string",
  instanceId = "string"
)
}
}

