% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_deployment_target}
\alias{get_deployment_target}
\title{Returns information about a deployment target}
\usage{
get_deployment_target(deploymentId = NULL, targetId = NULL)
}
\arguments{
\item{deploymentId}{The unique ID of a deployment.}

\item{targetId}{The unique ID of a deployment target.}
}
\description{
Returns information about a deployment target.
}
\section{Accepted Parameters}{
\preformatted{get_deployment_target(
  deploymentId = "string",
  targetId = "string"
)
}
}

