% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_on_premises_instance}
\alias{get_on_premises_instance}
\title{Gets information about an on-premises instance}
\usage{
get_on_premises_instance(instanceName)
}
\arguments{
\item{instanceName}{The name of the on-premises instance about which to get information.}
}
\description{
Gets information about an on-premises instance.
}
\examples{

}
