% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{list_applications}
\alias{list_applications}
\title{Lists the applications registered with the applicable IAM user or AWS account}
\usage{
list_applications(nextToken = NULL)
}
\arguments{
\item{nextToken}{An identifier returned from the previous list applications call. It can be used to return the next set of applications in the list.}
}
\description{
Lists the applications registered with the applicable IAM user or AWS account.
}
\examples{

}
