% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{list_deployment_configs}
\alias{list_deployment_configs}
\title{Lists the deployment configurations with the applicable IAM user or AWS account}
\usage{
list_deployment_configs(nextToken = NULL)
}
\arguments{
\item{nextToken}{An identifier returned from the previous list deployment configurations call. It can be used to return the next set of deployment configurations in the list.}
}
\description{
Lists the deployment configurations with the applicable IAM user or AWS account.
}
\examples{

}
