% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{list_deployment_instances}
\alias{list_deployment_instances}
\title{The newer BatchGetDeploymentTargets should be used instead because it works with all compute types}
\usage{
list_deployment_instances(
  deploymentId = "string",
  nextToken = "string",
  instanceStatusFilter = list(
    "Pending"|"InProgress"|"Succeeded"|"Failed"|"Skipped"|"Unknown"|"Ready"
  ),
  instanceTypeFilter = list(
    "Blue"|"Green"
  )
)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{nextToken}{An identifier returned from the previous list deployment instances call. It can be used to return the next set of deployment instances in the list.}

\item{instanceStatusFilter}{A subset of instances to list by status:
\itemize{
\item Pending: Include those instance with pending deployments.
\item InProgress: Include those instance where deployments are still in progress.
\item Succeeded: Include those instances with successful deployments.
\item Failed: Include those instance with failed deployments.
\item Skipped: Include those instance with skipped deployments.
\item Unknown: Include those instance with deployments in an unknown state.
}}

\item{instanceTypeFilter}{The set of instances in a blue/green deployment, either those in the original environment ("BLUE") or those in the replacement environment ("GREEN"), for which you want to view instance information.}
}
\description{
The newer BatchGetDeploymentTargets should be used instead because it works with all compute types. \code{ListDeploymentInstances} throws an exception if it is used with a compute platform other than EC2/On-premises or AWS Lambda.
}
\details{
Lists the instance for a deployment associated with the applicable IAM user or AWS account.
}
\examples{

}
