% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{list_deployment_targets}
\alias{list_deployment_targets}
\title{Returns an array of target IDs that are associated a deployment}
\usage{
list_deployment_targets(
  deploymentId = "string",
  nextToken = "string",
  targetFilters = list(
    list(
      "string"
    )
  )
)
}
\arguments{
\item{deploymentId}{The unique ID of a deployment.}

\item{nextToken}{A token identifier returned from the previous \code{ListDeploymentTargets} call. It can be used to return the next set of deployment targets in the list.}

\item{targetFilters}{A key used to filter the returned targets.}
}
\description{
Returns an array of target IDs that are associated a deployment.
}
\examples{

}
