% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{list_git_hub_account_token_names}
\alias{list_git_hub_account_token_names}
\title{Lists the names of stored connections to GitHub accounts}
\usage{
list_git_hub_account_token_names(nextToken = NULL)
}
\arguments{
\item{nextToken}{An identifier returned from the previous ListGitHubAccountTokenNames call. It can be used to return the next set of names in the list.}
}
\description{
Lists the names of stored connections to GitHub accounts.
}
\section{Accepted Parameters}{
\preformatted{list_git_hub_account_token_names(
  nextToken = "string"
)
}
}

