% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_package.R
\docType{package}
\name{paws.codedeploy-package}
\alias{paws.codedeploy}
\alias{paws.codedeploy-package}
\title{paws.codedeploy: AWS CodeDeploy}
\description{
AWS CodeDeploy is a deployment service that
    automates application deployments to Amazon EC2 instances, on-premises
    instances running in your own facility, or serverless AWS Lambda
    functions.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
