% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{register_application_revision}
\alias{register_application_revision}
\title{Registers with AWS CodeDeploy a revision for the specified application}
\usage{
register_application_revision(applicationName, revision, description = NULL)
}
\arguments{
\item{applicationName}{The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}

\item{revision}{Information about the application revision to register, including type and location.}

\item{description}{A comment about the revision.}
}
\description{
Registers with AWS CodeDeploy a revision for the specified application.
}
\examples{

}
