% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{stop_deployment}
\alias{stop_deployment}
\title{Attempts to stop an ongoing deployment}
\usage{
stop_deployment(deploymentId, autoRollbackEnabled = NULL)
}
\arguments{
\item{deploymentId}{The unique ID of a deployment.}

\item{autoRollbackEnabled}{Indicates, when a deployment is stopped, whether instances that have been updated should be rolled back to the previous version of the application revision.}
}
\description{
Attempts to stop an ongoing deployment.
}
\examples{

}
