% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{update_application}
\alias{update_application}
\title{Changes the name of an application}
\usage{
update_application(applicationName = NULL, newApplicationName = NULL)
}
\arguments{
\item{applicationName}{The current name of the application you want to change.}

\item{newApplicationName}{The new name to give the application.}
}
\description{
Changes the name of an application.
}
\examples{

}
