% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{batch_get_application_revisions}
\alias{batch_get_application_revisions}
\title{Gets information about one or more application revisions}
\usage{
batch_get_application_revisions(applicationName, revisions)
}
\arguments{
\item{applicationName}{The name of an AWS CodeDeploy application about which to get revision information.}

\item{revisions}{Information to get about the application revisions, including type and location.}
}
\description{
Gets information about one or more application revisions.
}
\examples{

}
