% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{batch_get_deployment_instances}
\alias{batch_get_deployment_instances}
\title{Gets information about one or more instance that are part of a deployment group}
\usage{
batch_get_deployment_instances(deploymentId, instanceIds)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{instanceIds}{[required] The unique IDs of instances in the deployment group.}
}
\description{
Gets information about one or more instance that are part of a deployment group.
}
\section{Accepted Parameters}{
\preformatted{batch_get_deployment_instances(
  deploymentId = "string",
  instanceIds = list(
    "string"
  )
)
}
}

