% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_deployment}
\alias{get_deployment}
\title{Gets information about a deployment}
\usage{
get_deployment(deploymentId)
}
\arguments{
\item{deploymentId}{A deployment ID associated with the applicable IAM user or AWS account.}
}
\description{
Gets information about a deployment.
}
\examples{

}
