% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{get_deployment_group}
\alias{get_deployment_group}
\title{Gets information about a deployment group}
\usage{
get_deployment_group(applicationName, deploymentGroupName)
}
\arguments{
\item{applicationName}{The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}

\item{deploymentGroupName}{The name of an existing deployment group for the specified application.}
}
\description{
Gets information about a deployment group.
}
\examples{

}
