% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{list_deployment_groups}
\alias{list_deployment_groups}
\title{Lists the deployment groups for an application registered with the applicable IAM user or AWS account}
\usage{
list_deployment_groups(
  applicationName = "string",
  nextToken = "string"
)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.}

\item{nextToken}{An identifier returned from the previous list deployment groups call. It can be used to return the next set of deployment groups in the list.}
}
\description{
Lists the deployment groups for an application registered with the applicable IAM user or AWS account.
}
\examples{

}
