% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{register_application_revision}
\alias{register_application_revision}
\title{Registers with AWS CodeDeploy a revision for the specified application}
\usage{
register_application_revision(applicationName, description = NULL,
  revision)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the IAM user or AWS account.}

\item{description}{A comment about the revision.}

\item{revision}{[required] Information about the application revision to register, including type and location.}
}
\description{
Registers with AWS CodeDeploy a revision for the specified application.
}
\section{Accepted Parameters}{
\preformatted{register_application_revision(
  applicationName = "string",
  description = "string",
  revision = list(
    revisionType = "S3"|"GitHub"|"String"|"AppSpecContent",
    s3Location = list(
      bucket = "string",
      key = "string",
      bundleType = "tar"|"tgz"|"zip"|"YAML"|"JSON",
      version = "string",
      eTag = "string"
    ),
    gitHubLocation = list(
      repository = "string",
      commitId = "string"
    ),
    string = list(
      content = "string",
      sha256 = "string"
    ),
    appSpecContent = list(
      content = "string",
      sha256 = "string"
    )
  )
)
}
}

