% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codedeploy_operations.R
\name{skip_wait_time_for_instance_termination}
\alias{skip_wait_time_for_instance_termination}
\title{In a blue/green deployment, overrides any specified wait time and starts terminating instances immediately after the traffic routing is completed}
\usage{
skip_wait_time_for_instance_termination(
  deploymentId = "string"
)
}
\arguments{
\item{deploymentId}{The unique ID of a blue/green deployment for which you want to skip the instance termination wait time.}
}
\description{
In a blue/green deployment, overrides any specified wait time and starts terminating instances immediately after the traffic routing is completed.
}
\examples{

}
